# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LinksUpdate": {
        "description": "The full list, in display order.",
        "properties": {
          "links": {
            "items": {
              "$ref": "#/components/schemas/ProfileLinkIn"
            },
            "maxItems": 8,
            "title": "Links",
            "type": "array"
          }
        },
        "required": [
          "links"
        ],
        "title": "LinksUpdate",
        "type": "object"
      },
...
      "Me": {
        "description": "The signed-in account, as the web app needs it on every page.",
        "properties": {
          "accent_color": {
            "enum": [
              "cyan",
              "violet",
              "lime",
              "amber",
              "magenta",
              "coral"
            ],
            "title": "Accent Color",
            "type": "string"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "github_login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Login"
          },
          "has_password": {
            "title": "Has Password",
            "type": "boolean"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "id",
          "email",
          "email_verified",
          "username",
          "display_name",
          "headline",
          "accent_color",
          "avatar_url",
          "has_password",
          "github_login",
          "created_at"
        ],
        "title": "Me",
        "type": "object"
      },
...
      "OnboardingRequest": {
        "properties": {
          "display_name": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "headline": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "display_name"
        ],
        "title": "OnboardingRequest",
        "type": "object"
      },
...
      "ProfileLinkIn": {
        "properties": {
          "kind": {
            "enum": [
              "github",
              "gitlab",
              "linkedin",
              "x",
              "mastodon",
              "bluesky",
              "website",
              "blog",
              "youtube",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          },
          "url": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "url"
        ],
        "title": "ProfileLinkIn",
        "type": "object"
      },
      "ProfileLinkOut": {
        "properties": {
          "kind": {
            "enum": [
              "github",
              "gitlab",
              "linkedin",
              "x",
              "mastodon",
              "bluesky",
              "website",
              "blog",
              "youtube",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "url"
        ],
        "title": "ProfileLinkOut",
        "type": "object"
      },
      "ProfileSettings": {
        "properties": {
          "accent_color": {
            "enum": [
              "cyan",
              "violet",
              "lime",
              "amber",
              "magenta",
              "coral"
            ],
            "title": "Accent Color",
            "type": "string"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "banner_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banner Url"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/ProfileLinkOut"
            },
            "title": "Links",
            "type": "array"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "title": "Tags",
            "type": "array"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          }
        },
        "required": [
          "display_name",
          "headline",
          "bio",
          "location",
          "website",
          "accent_color",
          "avatar_url",
          "banner_url",
          "links",
          "tags"
        ],
        "title": "ProfileSettings",
        "type": "object"
      },
      "ProfileUpdate": {
        "description": "Fields left out are unchanged; fields sent as null or blank are cleared.",
        "properties": {
          "accent_color": {
            "anyOf": [
              {
                "enum": [
                  "cyan",
                  "violet",
                  "lime",
                  "amber",
                  "magenta",
                  "coral"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accent Color"
          },
          "bio": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 50,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "headline": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "location": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "website": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          }
        },
        "title": "ProfileUpdate",
        "type": "object"
      },
      "PublicProfile": {
        "properties": {
          "accent_color": {
            "enum": [
              "cyan",
              "violet",
              "lime",
              "amber",
              "magenta",
              "coral"
            ],
            "title": "Accent Color",
            "type": "string"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "banner_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banner Url"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "friend_count": {
            "title": "Friend Count",
            "type": "integer"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "joined_at": {
            "format": "date-time",
            "title": "Joined At",
            "type": "string"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/ProfileLinkOut"
            },
            "title": "Links",
            "type": "array"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "relationship": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Relationship"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "title": "Tags",
            "type": "array"
          },
          "username": {
            "title": "Username",
            "type": "string"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          }
        },
        "required": [
          "username",
          "display_name",
          "headline",
          "bio",
          "location",
          "website",
          "accent_color",
          "avatar_url",
          "banner_url",
          "links",
          "tags",
          "friend_count",
          "relationship",
          "joined_at"
        ],
        "title": "PublicProfile",
        "type": "object"
      },
      "Relationship": {
        "description": "How the signed-in person relates to someone. `request_id` is set while a request is open.",
        "properties": {
          "request_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "enum": [
              "self",
              "none",
              "friends",
              "request_sent",
              "request_received",
              "blocked"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "Relationship",
        "type": "object"
      },
...
      "TagOut": {
        "properties": {
          "kind": {
            "enum": [
              "language",
              "tool",
              "topic"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "kind"
        ],
        "title": "TagOut",
        "type": "object"
      },
...
      "UsernameAvailability": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "problem": {
            "anyOf": [
              {
                "enum": [
                  "invalid",
                  "reserved",
                  "taken"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "available",
          "problem"
        ],
        "title": "UsernameAvailability",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Lantern Grid API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me/links": {
      "put": {
        "operationId": "setMyLinks",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinksUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileLinkOut"
                  },
                  "title": "Response Setmylinks",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set My Links",
        "tags": [
          "users"
        ]
      }
    },
...
    "/v1/me/onboarding": {
      "post": {
        "operationId": "completeOnboarding",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Onboarding",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/me/profile": {
      "get": {
        "operationId": "getMyProfile",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSettings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get My Profile",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "updateMyProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update My Profile",
        "tags": [
          "users"
        ]
      }
    },
...
    "/v1/usernames/{username}": {
      "get": {
        "operationId": "checkUsername",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsernameAvailability"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Username",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/{username}": {
      "get": {
        "operationId": "getPublicProfile",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public Profile",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}